requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
pypdf>=4.0.0
faiss-cpu>=1.7.4
spacy>=3.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
